+++ routes/api.php
@@ -13,6 +13,7 @@ use App\Http\Controllers\OauthController;
 use App\Http\Controllers\ReportController;
 use App\Http\Controllers\SensorController;
 use App\Http\Controllers\SyncController;
+use App\Http\Controllers\CameraController;
 use Illuminate\Support\Facades\Route;
 
 Route::post('/login', [AuthController::class, 'login']);
@@ -77,38 +78,38 @@ Route::post('checkplayback', [MapController::class, 'checkplayback']);
 Route::post('exportais', [ExportController::class, 'exportais']);
 
 //camera
-Route::post('camzoomminus', [HelperController::class, 'camzoomminus']);
-Route::post('camzoomminuscon', [HelperController::class, 'camzoomminuscon']);
-Route::post('camup', [HelperController::class, 'camup']);
-Route::post('camupcon', [HelperController::class, 'camupcon']);
-Route::post('camzoomplus', [HelperController::class, 'camzoomplus']);
-Route::post('camzoompluscon', [HelperController::class, 'camzoompluscon']);
-Route::post('camleft', [HelperController::class, 'camleft']);
-Route::post('camleftcon', [HelperController::class, 'camleftcon']);
-Route::post('camleftup', [HelperController::class, 'camleftup']);
-Route::post('camleftupcon', [HelperController::class, 'camleftupcon']);
-Route::post('camright', [HelperController::class, 'camright']);
-Route::post('camrightup', [HelperController::class, 'camrightup']);
-Route::post('camdown', [HelperController::class, 'camdown']);
-Route::post('camautopan', [HelperController::class, 'camautopan']);
-Route::post('camautopanstop', [HelperController::class, 'camautopanstop']);
-Route::post('camstop', [HelperController::class, 'camstop']);
-Route::post('camfocusmin', [HelperController::class, 'camfocusmin']);
-Route::post('camfocusplus', [HelperController::class, 'camfocusplus']);
-Route::post('camfocusstop', [HelperController::class, 'camfocusstop']);
-Route::post('camleftdown', [HelperController::class, 'camleftdown']);
-Route::post('camrightdown', [HelperController::class, 'camrightdown']);
-Route::post('camirismin', [HelperController::class, 'camirismin']);
-Route::post('camirisplus', [HelperController::class, 'camirisplus']);
-Route::post('camirisstop', [HelperController::class, 'camirisstop']);
-Route::post('camwiper', [HelperController::class, 'camwiper']);
-Route::post('camstopwiper', [HelperController::class, 'camstopwiper']);
-Route::post('cammenu', [HelperController::class, 'cammenu']);
-Route::post('camstopzoom', [HelperController::class, 'camstopzoom']);
+Route::post('camzoomminus', [CameraController::class, 'camzoomminus']);
+Route::post('camzoomminuscon', [CameraController::class, 'camzoomminuscon']);
+Route::post('camup', [CameraController::class, 'camup']);
+Route::post('camupcon', [CameraController::class, 'camupcon']);
+Route::post('camzoomplus', [CameraController::class, 'camzoomplus']);
+Route::post('camzoompluscon', [CameraController::class, 'camzoompluscon']);
+Route::post('camleft', [CameraController::class, 'camleft']);
+Route::post('camleftcon', [CameraController::class, 'camleftcon']);
+Route::post('camleftup', [CameraController::class, 'camleftup']);
+Route::post('camleftupcon', [CameraController::class, 'camleftupcon']);
+Route::post('camright', [CameraController::class, 'camright']);
+Route::post('camrightup', [CameraController::class, 'camrightup']);
+Route::post('camdown', [CameraController::class, 'camdown']);
+Route::post('camautopan', [CameraController::class, 'camautopan']);
+Route::post('camautopanstop', [CameraController::class, 'camautopanstop']);
+Route::post('camstop', [CameraController::class, 'camstop']);
+Route::post('camfocusmin', [CameraController::class, 'camfocusmin']);
+Route::post('camfocusplus', [CameraController::class, 'camfocusplus']);
+Route::post('camfocusstop', [CameraController::class, 'camfocusstop']);
+Route::post('camleftdown', [CameraController::class, 'camleftdown']);
+Route::post('camrightdown', [CameraController::class, 'camrightdown']);
+Route::post('camirismin', [CameraController::class, 'camirismin']);
+Route::post('camirisplus', [CameraController::class, 'camirisplus']);
+Route::post('camirisstop', [CameraController::class, 'camirisstop']);
+Route::post('camwiper', [CameraController::class, 'camwiper']);
+Route::post('camstopwiper', [CameraController::class, 'camstopwiper']);
+Route::post('cammenu', [CameraController::class, 'cammenu']);
+Route::post('camstopzoom', [CameraController::class, 'camstopzoom']);
 
-Route::post('movebylatlng', [HelperController::class, 'movebylatlng']);
-Route::post('camset', [HelperController::class, 'camset']);
-Route::post('camcall', [HelperController::class, 'camcall']);
+Route::post('movebylatlng', [CameraController::class, 'movebylatlng']);
+Route::post('camset', [CameraController::class, 'camset']);
+Route::post('camcall', [CameraController::class, 'camcall']);
 
 
 Route::post('updateradarname', [HelperController::class, 'updateradarname']);
